Fixed problems with ft_strlen_ignore

+++ srcs/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: mialbert <mialbert@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/15 21:19:34 by mialbert          #+#    #+#             */
-/*   Updated: 2021/12/29 00:42:21 by mialbert         ###   ########.fr       */
+/*   Updated: 2022/07/17 21:14:05 by mialbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "../includes/libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
